+++ web/modules/custom/xmt_trust_ui/src/Controller/TrustFeedController.php
@@ -69,6 +69,7 @@ class TrustFeedController extends ControllerBase {
       'official' => Link::fromTextAndUrl($this->t('Official (L1)'), Url::fromRoute('xmt_trust_ui.feed_official'))->toRenderable(),
       'enterprise' => Link::fromTextAndUrl($this->t('Enterprise (L2)'), Url::fromRoute('xmt_trust_ui.feed_enterprise'))->toRenderable(),
       'aggregate' => Link::fromTextAndUrl($this->t('Aggregate (L0)'), Url::fromRoute('xmt_trust_ui.feed_aggregate'))->toRenderable(),
+      'publishers' => Link::fromTextAndUrl($this->t('Publishers'), Url::fromRoute('xmt_trust_ui.publishers_directory'))->toRenderable(),
     ];
 
     $formats = [